feat(harmonia): humanize generated property labels + task-form Details-card layout
Two presentation fixes to the generated Harmonia UI, reported together.

1. Humanized property labels. The label catalog value for a property defaulted to the
   raw PascalCase name (parameterUtils widgetLabel default -> generateUtils translate
   action), so list columns and the detail/document view showed "PaymentMethod" /
   "TaxEventDate" instead of "Payment Method" / "Tax Event Date" - while entity
   labels were already humanized. humanizeName moves to parameterUtils (the shared util
   module generateUtils already imports from), is exported, and now backs the
   widgetLabel default; an explicitly authored widgetLabel still wins. Affects every
   generated app (Angular + Harmonia); flows to the label catalog and the form/list
   templates alike.

2. Task-form Details-card layout. A BPM task form's read-only fields rendered as
   label-beside-value hbox rows; they now render as the entity detail / document view's
   headerless x-h-card of vbox detail cells (muted text-xs label above a text-sm value),
   so the Inbox task form matches the entity Details layout.

Verified live: regenerated kf-mod-sales-invoices - the served label catalog humanizes
(Payment Method / Tax Event Date / Bank Account) and the ApproveSalesInvoice task form
renders the x-h-card + vbox gap-0 detail cells.
